extract out common functionalities
