+++ src/main/java/com/twilio/oai/AbstractTwilioGoGenerator.java
@@ -1,11 +1,16 @@
 package com.twilio.oai;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import org.openapitools.codegen.CodegenConstants;
 import org.openapitools.codegen.CodegenProperty;
 import org.openapitools.codegen.languages.GoClientCodegen;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.parameters.Parameter;
+import org.openapitools.codegen.CodegenParameter;
 
 public abstract class AbstractTwilioGoGenerator extends GoClientCodegen {
 
@@ -33,6 +38,33 @@ public abstract class AbstractTwilioGoGenerator extends GoClientCodegen {
         }
 
         return super.createMapping(key, value);
+    }
+
+    @Override
+    public void postProcessParameter(final CodegenParameter parameter) {
+        // Make sure required non-path params get into the options block.
+        parameter.required = parameter.isPathParam;
+
+        if (parameter.paramName.equals("PathAccountSid")) {
+            parameter.required = false;
+            parameter.vendorExtensions.put("x-is-account-sid", true);
+        }
+    }
+
+    @Override
+    public void processOpenAPI(final OpenAPI openAPI) {
+        openAPI.getPaths().forEach((name, path) -> path.readOperations().forEach(operation -> {
+            List<Parameter> parameters = operation.getParameters();
+            if (parameters != null) {
+                for (Parameter p : parameters) {
+                    String in = p.getIn();
+                    String paramName = p.getName();
+                    if (in.equals("path") && paramName.equals("AccountSid")) {
+                        p.setName("PathAccountSid");
+                    }
+                }
+            }
+        }));
     }
 
     @Override
